+++ lib/home/app_data.dart
@@ -34,6 +34,7 @@ import '../kccx/kccx_page.dart';
 import '../qxfacx/qxfacx_page.dart';
 import '../wspj/wspj_page.dart';
 import '../mail/mail_page.dart';
+import '../leader_mail/leader_mail_page.dart';
 
 /// 应用分类
 enum AppCategory { jiaowu, service, news }
@@ -134,6 +135,8 @@ final List<AppEntry> allApps = [
   AppEntry(icon: Icons.mail_rounded, name: '邮件系统', category: AppCategory.service,
     requiresLogin: true,
     pageBuilder: (ctx, c, uid) => MailPage(client: c)),
+  AppEntry(icon: Icons.mark_email_unread_rounded, name: '领导信箱', category: AppCategory.service,
+    pageBuilder: (ctx, c, uid) => const LeaderMailPage()),
   AppEntry(icon: Icons.forum_rounded, name: 'QQ频道', category: AppCategory.service,
     pageBuilder: (ctx, c, uid) => const WebViewPage(
       url: 'https://pd.qq.com/s/bq4dam2kg',